+++ diive/pkgs/qaqc/meteoscreening.py
@@ -275,7 +275,7 @@ class StepwiseMeteoScreeningDb:
 
             # Heatmaps
             kwargs_heatmap = dict(cb_labelsize=10, axlabels_fontsize=10, ticks_labelsize=10,
-                                  minyticks=3, maxyticks=99)
+                                  minticks=3, maxticks=99)
             HeatmapDateTime(ax=ax_heatmap_hires_before, series=series_orig, **kwargs_heatmap).plot()
             HeatmapDateTime(ax=ax_heatmap_resampled_after, series=series_resampled, **kwargs_heatmap).plot()
 
@@ -792,13 +792,13 @@ def _example():
 
     TESTDIR = Path(r"F:\TMP")
 
-    SITE = 'ch-hon'
-    SITE_LAT = 47.41887  # CH-HON
-    SITE_LON = 8.491318  # CH-HON
-    FIELDS = ['TA_T1_4_2']
+    SITE = 'ch-dav'
+    SITE_LAT = 46.815333  # CH-DAV
+    SITE_LON = 9.855972  # CH-DAV
+    FIELDS = ['TA_NABEL_T1_35_1']
     MEASUREMENT = 'TA'
-    START = '2025-02-01 00:00:01'
-    STOP = '2025-03-01 00:00:01'
+    START = '1997-01-01 00:00:01'
+    STOP = '2000-01-01 00:00:01'
     RESAMPLING_AGG = 'mean'
     # RESAMPLING_AGG = 'sum'
     DATA_VERSION = 'raw'
@@ -824,9 +824,9 @@ def _example():
     print(data_simple)
     print(data_detailed)
     print(assigned_measurements)
-    import matplotlib.pyplot as plt
-    data_simple.plot()
-    plt.show()
+    # import matplotlib.pyplot as plt
+    # data_simple.plot()
+    # plt.show()
 
     # print(f"Data available for: {data_detailed.keys()}\n")
     # vars_not_available = [v for v in FIELDS if v not in data_detailed.keys()]
@@ -865,13 +865,13 @@ def _example():
                                     site_lat=SITE_LAT,
                                     site_lon=SITE_LON,
                                     utc_offset=TIMEZONE_OFFSET_TO_UTC_HOURS)
-    mscr.showplot_orig()
+    # mscr.showplot_orig()
     mscr.showplot_cleaned()
 
     # -----------------
     # OUTLIER DETECTION
     # -----------------
-    mscr.start_outlier_detection()  # If needed
+    # mscr.start_outlier_detection()  # If needed
     # REMOVE_DATES = [
     #     ['2010-12-24 07:15:00', '2011-01-01 00:00:15'],  # Remove time range
     #     # '2022-08-23 11:45:00',  # Remove data point
@@ -879,16 +879,16 @@ def _example():
     # ]
     # mscr.flag_manualremoval_test(remove_dates=REMOVE_DATES, showplot=True, verbose=True)
     # mscr.flag_outliers_zscore_dtnt_test(thres_zscore=4, showplot=True, verbose=True, repeat=True)
-    mscr.flag_outliers_hampel_test(window_length=48 * 7, n_sigma=5.5, showplot=True, verbose=True, repeat=True)
-    mscr.flag_outliers_hampel_dtnt_test(window_length=48 * 7, n_sigma_dt=5.5, n_sigma_nt=4.5, showplot=True,
-                                        verbose=True, repeat=True)
-    mscr.addflag()
-    mscr.flag_outliers_trim_low_test(trim_daytime=False, trim_nighttime=True, lower_limit=10,
-                                     showplot=True, verbose=True)
-    mscr.addflag()
+    # mscr.flag_outliers_hampel_test(window_length=48 * 7, n_sigma=5.5, showplot=True, verbose=True, repeat=True)
+    # mscr.flag_outliers_hampel_dtnt_test(window_length=48 * 7, n_sigma_dt=5.5, n_sigma_nt=4.5, showplot=True,
+    #                                     verbose=True, repeat=True)
+    # mscr.addflag()
+    # mscr.flag_outliers_trim_low_test(trim_daytime=False, trim_nighttime=True, lower_limit=10,
+    #                                  showplot=True, verbose=True)
+    # mscr.addflag()
     # mscr.flag_outliers_zscore_rolling_test(thres_zscore=3, winsize=48, showplot=True, verbose=True, repeat=True)
     # mscr.flag_outliers_localsd_test(n_sd=2.5, winsize=24, showplot=True, verbose=True, repeat=False)
-    mscr.flag_outliers_localsd_test(n_sd=2.5, winsize=24, showplot=True, verbose=True, repeat=False)
+    # mscr.flag_outliers_localsd_test(n_sd=2.5, winsize=24, showplot=True, verbose=True, repeat=False)
     # mscr.flag_outliers_increments_zcore_test(thres_zscore=9, showplot=True, verbose=True, repeat=True)
     # mscr.flag_outliers_zscore_test(thres_zscore=15, showplot=True, verbose=True, repeat=True)
     # mscr.flag_outliers_lof_dtnt_test(n_neighbors=3, contamination=0.00001, showplot=True,
@@ -916,13 +916,13 @@ def _example():
     # # -----------
     # mscr.showplot_orig()
     # mscr.showplot_cleaned()
-    mscr.correction_remove_radiation_zero_offset()  # Remove radiation zero offset
+    # mscr.correction_remove_radiation_zero_offset()  # Remove radiation zero offset
     # mscr.correction_remove_relativehumidity_offset()  # Remove relative humidity offset
     # mscr.correction_setto_max_threshold(threshold=30)  # Set to max threshold
     # mscr.correction_setto_min_threshold(threshold=-5)  # Set to min threshold
     # DATES = [['2022-04-01', '2022-04-05'], ['2022-09-05', '2022-09-07']]
     # mscr.correction_setto_value(dates=DATES, value=3.7, verbose=1)  # Set to value
-    mscr.correction_set_exact_value_to_missing(values=[0])
+    # mscr.correction_set_exact_value_to_missing(values=[0])
 
     # # -----------
     # # ANALYSES
@@ -934,7 +934,7 @@ def _example():
     # RESAMPLING
     # -----------
     mscr.resample(to_freqstr=RESAMPLING_FREQ, agg=RESAMPLING_AGG, mincounts_perc=.25)
-    # mscr.showplot_resampled()
+    mscr.showplot_resampled()
     for v in mscr.resampled_detailed.keys():
         print(f"{'-' * 20}\n{v}")
         _checkfreq = DetectFrequency(index=mscr.resampled_detailed[v].index, verbose=True).get()
@@ -950,20 +950,20 @@ def _example():
     #     dc = DielCycle(series=series)
     #     dc.plot(ax=None, each_month=True, legend_n_col=2)
 
-    # -----------
-    # UPLOAD
-    # -----------
-    from dbc_influxdb import dbcInflux
-    dbc = dbcInflux(dirconf=DIRCONF)
-    for v in mscr.resampled_detailed.keys():
-        # mscr.resampled_detailed[v][v] = mscr.resampled_detailed[v][v].multiply(999)
-        # mscr.resampled_detailed[v]['hpos'] = '999'
-        m = assigned_measurements[v]
-        dbc.upload_singlevar(to_bucket=BUCKET_PROCESSED,
-                             to_measurement=m,
-                             var_df=mscr.resampled_detailed[v],
-                             timezone_offset_to_utc_hours=TIMEZONE_OFFSET_TO_UTC_HOURS,
-                             delete_from_db_before_upload=True)
+    # # -----------
+    # # UPLOAD
+    # # -----------
+    # from dbc_influxdb import dbcInflux
+    # dbc = dbcInflux(dirconf=DIRCONF)
+    # for v in mscr.resampled_detailed.keys():
+    #     # mscr.resampled_detailed[v][v] = mscr.resampled_detailed[v][v].multiply(999)
+    #     # mscr.resampled_detailed[v]['hpos'] = '999'
+    #     m = assigned_measurements[v]
+    #     dbc.upload_singlevar(to_bucket=BUCKET_PROCESSED,
+    #                          to_measurement=m,
+    #                          var_df=mscr.resampled_detailed[v],
+    #                          timezone_offset_to_utc_hours=TIMEZONE_OFFSET_TO_UTC_HOURS,
+    #                          delete_from_db_before_upload=True)
 
     # -----------
     # DOWNLOAD